requirements.txt

aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via discord-py
aiosignal==1.3.2
    # via aiohttp
attrs==24.3.0
    # via aiohttp
blinker==1.9.0
    # via
    #   -r requirements.in
    #   flask
click==8.1.8
    # via
    #   -r requirements.in
    #   click-aliases
    #   click-default-group
    #   click-help-colors
    #   flask
click-aliases==1.0.5
    # via -r requirements.in
click-default-group==1.2.4
    # via -r requirements.in
click-help-colors==0.9.4
    # via -r requirements.in
colorama==0.4.6
    # via
    #   -r requirements.in
    #   click
discord-py==2.4.0
    # via -r requirements.in
flask==3.1.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.5
    # via
    #   -r requirements.in
    #   flask
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.0.1
    # via -r requirements.in
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   flask
yarl==1.18.3
    # via aiohttp
gunicorn==21.2.0
pymongo==4.6.1
dnspython
aiocron
PyNaCl
psutil
motor
pytz
matplotlib
